+++ tests/Unit/FlashesTest.php
@@ -3,10 +3,10 @@
 namespace Tests\Unit;
 
 it('should render flashes component', function () {
-    expect('<ul></ul>')->toBeRenderOf('<x-ui-flashes/>');
+    expect('')->toBeRenderOf('<x-ui-flashes/>');
 
-    $items = "['success' => 'Success', 'failure' => 'Failure']";
+    $messages = "['success' => 'Success', 'failure' => 'Failure']";
     expect('<ul><li>Success</li><li>Failure</li></ul>')
-        ->toBeRenderOf('<x-ui-flashes :list="'.$items.'"/>')
-        ->toBeRenderOf('<x-ui::flashes :list="'.$items.'"/>');
+        ->toBeRenderOf('<x-ui-flashes :messages="'.$messages.'"/>')
+        ->toBeRenderOf('<x-ui::flashes :messages="'.$messages.'"/>');
 });